some stuff, experimental bitsolution

+++ network_project/client/request/command/utils/scan.h
@@ -9,6 +9,6 @@
 #define SCAN_COMPL 2
 
 #define check_ascii(B) (B >= 0x00 && B <= 0x7F)
-int8_t scan_driver(char *scanner, uint32_t size_buffer, char *message);
+int32_t scan_driver(char *scanner, uint32_t size_buffer, char *message);
 
 #endif
